# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AlertChannel": {
        "description": "One delivery channel of an alert rule (`alert_rule.channels` element).",
        "oneOf": [
          {
            "description": "Email delivery (instant or digest).",
            "properties": {
              "to": {
                "description": "Destination address.",
                "type": "string"
              },
              "type": {
                "enum": [
                  "email"
                ],
                "type": "string"
              }
            },
            "required": [
              "to",
              "type"
            ],
            "type": "object"
          },
          {
            "description": "HMAC-SHA256-signed webhook (design §6.4): the dispatcher POSTs JSON\nand signs the exact body with `secret` in the\n`X-Govfolio-Signature: sha256=<hex>` header.",
            "properties": {
              "secret": {
                "description": "Per-rule shared secret for the HMAC signature.",
                "type": "string"
              },
              "type": {
                "enum": [
                  "webhook"
                ],
                "type": "string"
              },
              "url": {
                "description": "POST target URL.",
                "type": "string"
              }
            },
            "required": [
              "url",
              "secret",
              "type"
            ],
            "type": "object"
          }
        ]
      },
      "AlertRule": {
        "description": "One alert rule.",
        "properties": {
          "active": {
            "description": "Inactive rules never match.",
            "type": "boolean"
          },
          "channels": {
            "description": "Delivery channels; at most one per channel type.",
            "items": {
              "$ref": "#/components/schemas/AlertChannel"
            },
            "type": "array"
          },
          "created_at": {
            "description": "When the rule was created.",
            "format": "date-time",
            "type": "string"
          },
          "digest": {
            "description": "Digest mode: matches accumulate and ship as one periodic summary\ninstead of instant per-record sends.",
            "type": "boolean"
          },
          "filter": {
            "$ref": "#/components/schemas/RecordFilter",
            "description": "Filter in the shared record grammar — identical to `/v1/records`\nquery parameters (design §6.3: one grammar, learned once)."
          },
          "id": {
            "description": "Rule ULID.",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "type": "string"
          },
          "updated_at": {
            "description": "When the rule was last updated.",
            "format": "date-time",
            "type": "string"
          },
          "user_id": {
            "description": "Owning account (`user_account.id` — always the authenticated caller).",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "filter",
          "channels",
          "digest",
          "active",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
      "AlertRuleSpec": {
        "description": "Create/replace body for an alert rule (`POST` and `PUT` share it). The\nowner is always the authenticated account — never caller-supplied.",
        "properties": {
          "active": {
            "description": "Whether the rule matches at all (defaults to active).",
            "type": "boolean"
          },
          "channels": {
            "description": "Delivery channels; at least one, at most one per channel type.",
            "items": {
              "$ref": "#/components/schemas/AlertChannel"
            },
            "type": "array"
          },
          "digest": {
            "description": "Digest mode (defaults to instant).",
            "type": "boolean"
          },
          "filter": {
            "$ref": "#/components/schemas/RecordFilter",
            "description": "Filter in the shared record grammar; unknown keys are rejected."
          }
        },
        "required": [
          "filter",
          "channels"
        ],
        "type": "object"
      },
...
      "AssetClass": {
        "description": "`asset_class` vocabulary (design §4.2 column comment).",
        "enum": [
          "equity",
          "bond",
          "fund",
          "option",
          "crypto",
          "commodity",
          "real_estate",
          "private",
          "other"
        ],
        "type": "string"
      },
...
      "ErrorBody": {
        "description": "The error envelope every non-2xx response carries.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "The error itself."
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "ErrorDetail": {
        "description": "Machine-readable error code plus a human-readable message.",
        "properties": {
          "code": {
            "description": "Stable machine-readable code (e.g. `invalid_cursor`, `not_found`).",
            "examples": [
              "invalid_cursor"
            ],
            "type": "string"
          },
          "message": {
            "description": "Human-readable explanation.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "InstrumentId": {
        "description": "ULID (26-char Crockford base32, time-sortable)",
        "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
        "type": "string"
      },
...
      "PoliticianId": {
        "description": "ULID (26-char Crockford base32, time-sortable)",
        "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
        "type": "string"
      },
...
      "RecordPage": {
        "description": "One page of records; `next_cursor` is the last item's ULID, `null` when\nthe listing is exhausted (design §6.1: cursor pagination on ULIDs).",
        "properties": {
          "items": {
            "description": "Records in ascending ULID (= insertion-time) order.",
            "items": {
              "$ref": "#/components/schemas/DisclosureRecord"
            },
            "type": "array"
          },
          "next_cursor": {
            "description": "Pass back as `cursor` to fetch the page after this one; `null` at the end.",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "RecordType": {
        "description": "`record_type` CHECK: the four observation types, one table (design D1).",
        "enum": [
          "transaction",
          "holding",
          "interest",
          "change_notification"
        ],
        "type": "string"
      },
...
      "VerificationState": {
        "description": "`verification_state` CHECK: two-stage publication states (design §4.2).",
        "enum": [
          "unverified",
          "verified",
          "corrected",
          "disputed"
        ],
        "type": "string"
      }
    }
  },
  "info": {
    "description": "Worldwide politician financial-disclosure tracking. Cursor pagination on ULIDs; consistent error envelope; verification_state on every record. Every successful GET carries a strong ETag (sha256 of the body); requests with a matching If-None-Match receive 304 Not Modified. Authentication: `Authorization: Bearer gfk_...` API keys (create via /v1/keys; hashes only at rest). Requests without a key are served at the free tier. Tiers control record freshness (free: filings we discovered less than 24 hours ago are not yet visible; pro/data: real time) and the daily request quota (counted per account; over quota is 429 quota_exceeded in the standard envelope). Anonymous traffic has a light per-IP rate backstop (429 rate_limited).",
    "title": "govfolio API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/alert-rules": {
      "get": {
        "description": "# Errors\n`401`/`403` without a pro/data key; `500` on backend failure.",
        "operationId": "list_alert_rules",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertRule"
                  },
                  "type": "array"
                }
              }
            },
            "description": "The caller's rules"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No valid API key"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Tier does not include alerts"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Lists the authenticated account's alert rules.",
        "tags": [
          "alert-rules"
        ]
      },
      "post": {
        "description": "# Errors\n`401`/`403` without a pro/data key; `400` on a body outside the\ncontracts; `500` on backend failure.",
        "operationId": "create_alert_rule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertRuleSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRule"
                }
              }
            },
            "description": "The created rule"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Body violates the filter/channel contracts"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No valid API key"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Tier does not include alerts"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Creates an alert rule owned by the authenticated account.",
        "tags": [
          "alert-rules"
        ]
      }
    },
    "/v1/alert-rules/{id}": {
      "delete": {
        "description": "# Errors\n`401`/`403` without a pro/data key; `404` for a rule that is unknown or\nnot the caller's; `500` on backend failure.",
        "operationId": "delete_alert_rule",
        "parameters": [
          {
            "description": "Rule ULID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No valid API key"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Tier does not include alerts"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unknown (or not owned) rule"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Deletes one of the authenticated account's alert rules (its delivery\nledger rows cascade with it).",
        "tags": [
          "alert-rules"
        ]
      },
      "put": {
        "description": "# Errors\n`401`/`403` without a pro/data key; `400` on a body outside the\ncontracts; `404` for a rule that is unknown or not the caller's; `500`\non backend failure.",
        "operationId": "update_alert_rule",
        "parameters": [
          {
            "description": "Rule ULID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertRuleSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRule"
                }
              }
            },
            "description": "The updated rule"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Body violates the filter/channel contracts"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No valid API key"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Tier does not include alerts"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unknown (or not owned) rule"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Replaces one of the authenticated account's alert rules.",
        "tags": [
          "alert-rules"
        ]
      }
    },
...
    "/v1/records": {
      "get": {
        "description": "# Errors\n`400` on a malformed cursor, limit or filter; `500` on backend failure —\nall in the consistent error envelope.",
        "operationId": "list_records",
        "parameters": [
          {
            "description": "Pagination cursor: the `id` of the last record on the previous page\n(from `next_cursor`). Pages begin strictly after it.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
              "type": "string"
            }
          },
          {
            "description": "Page size, `1..=200`; defaults to 50.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Only records filed under a regime of this jurisdiction\n(`jurisdiction.id`, e.g. `us`).",
            "in": "query",
            "name": "jurisdiction_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only records of this type.",
            "in": "query",
            "name": "record_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RecordType"
            }
          },
          {
            "description": "Only records of this asset class.",
            "in": "query",
            "name": "asset_class",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AssetClass"
            }
          },
          {
            "description": "Only records concerning this politician.",
            "in": "query",
            "name": "politician_id",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PoliticianId"
            }
          },
          {
            "description": "Only records resolved to this instrument (below-threshold matches are\nNULL and never match — invariant 3).",
            "in": "query",
            "name": "instrument_id",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/InstrumentId"
            }
          },
          {
            "description": "Only records with `event_date` on or after this date.",
            "in": "query",
            "name": "event_date_from",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "description": "Only records with `event_date` on or before this date.",
            "in": "query",
            "name": "event_date_to",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "description": "Only records whose declared value band can reach this amount or more\n(decimal string — invariant 7).",
            "in": "query",
            "name": "value_min",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only records whose declared value band can stay at this amount or less\n(decimal string — invariant 7).",
            "in": "query",
            "name": "value_max",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only records in this verification state.",
            "in": "query",
            "name": "verification_state",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/VerificationState"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordPage"
                }
              }
            },
            "description": "One page of disclosure records; every record carries verification_state"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Malformed cursor, limit or filter"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Lists disclosure records in ascending ULID order.",
        "tags": [
          "records"
        ]
      }
    },
...
  },
...
}